NMU for CVE-2024-50349 and CVE-2024-52006
I've prepared an NMU for git (versioned as 1:2.47.2-0.1) and
uploaded it to DELAYED/2. Please feel free to tell me if I
should delay it longer.

I'm not sure if you have concerns about the short delay but I
defintively strongly would prefer to have the update first exposed in
unstable, then go down to bookworm releasing the packages.

What I tried is to follow as close as possible your current pckaging
from a git clone merge the upstream v2.47.2 tag for the unstable
upload as this is a supported version.

Differently though for bookworm and cherry-picking there the commits
from the closest down version.

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>