debian/rules: recover build by removing overly strict check of license text
In 48af9ce49bca (debian: Link extra license files to common-licenses,
2016-10-31), we replaced the content of these license files from
contrib with symlinks to files from base-files and used "diff" to
confirm we are not accidentally shipping something out of date.  But
we didn't take into account that base-files would make some minor
improvements to the canonical files itself.

Those improvements are okay for us; we can continue to ship symlinks.

If the files were to diverge in a more significant way, then the
license auditing that already happens to keep debian/copyright up to
date would notice that.  So this change should be safe.

Thanks to Santiago Vila for the observation and suggested fix.

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>